+++ CHANGELOG.md
@@ -22,6 +22,13 @@ You can easily find changes by searching for `"### Changed"` in the in this mark
 # Releases
 Test/interim releases are not documented here.
 
+---
+
+## wip
+### Added
+- Transpiler - support narrowing casts.
+    - `b = (u8)my_i32;` --> `b = (uint8_t)my_i32;`
+
 --- 
 
 ## [0.2.1-alpha]